+++ sickrage/core/common.py
@@ -558,6 +558,7 @@ countryList = {'Australia': 'AU',
                'Canada': 'CA',
                'USA': 'US'
                }
+
 dateFormat = '%Y-%m-%d'
 dateTimeFormat = '%Y-%m-%d %H:%M:%S'
 timeFormat = '%A %I:%M %p'