+++ sickrage/providers/torrent/extratorrent.py
@@ -32,7 +32,7 @@ from sickrage.providers import TorrentProvider
 
 class ExtraTorrentProvider(TorrentProvider):
     def __init__(self):
-        super(ExtraTorrentProvider, self).__init__("ExtraTorrent",'extratorrent.cc')
+        super(ExtraTorrentProvider, self).__init__("ExtraTorrent",'extratorrent.cc', False)
 
         self.urls.update({
             'rss': '{base_url}/rss.xml'.format(base_url=self.urls['base_url'])