+++ sickrage/providers/torrent/hdspace.py
@@ -33,7 +33,7 @@ from sickrage.providers import TorrentProvider
 
 class HDSpaceProvider(TorrentProvider):
     def __init__(self):
-        super(HDSpaceProvider, self).__init__("HDSpace", 'hd-space.org')
+        super(HDSpaceProvider, self).__init__("HDSpace", 'hd-space.org', True)
 
         self.supportsBacklog = True