+++ sickrage/providers/torrent/scenetime.py
@@ -31,7 +31,7 @@ from sickrage.providers import TorrentProvider
 class SceneTimeProvider(TorrentProvider):
     def __init__(self):
 
-        super(SceneTimeProvider, self).__init__("SceneTime",'www.scenetime.com')
+        super(SceneTimeProvider, self).__init__("SceneTime",'www.scenetime.com', True)
 
         self.supportsBacklog = True