+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * ddad0b5 - 2016-08-30: Fixed several issues relating to NZB/Newznab providers and searches 
+- * dd3e9dc - 2016-08-30: Fixed several issues relating to NZB/Newznab providers and searches 
+- * 2d06621 - 2016-08-30: Fixed several issues relating to NZB/Newznab providers and searches 
 - * c45d6e5 - 2016-08-28: Added code to extract sizes of files from downloaded torrent files. Compares provider search result size against default quality sizes based on quality of result. 
 - * e2793c2 - 2016-08-27: Fixed issues with NZB searches and params 
 - * 9809565 - 2016-08-27: Updated to version 8.0.9